Added a mainmenu
Game mode not functionals yet

To deal with the increasing number of state (game/menu/etc), a statestack is introduced on which we can put/pop states. Only the top of the stack has "focus" (i.e. receive update/draw/etc)

Also, lua_mod is now global and defined in main